feat: 관광지 타입 option 태그 추가 & vue 파일 이동
- KaKaoMap.vue trip -> map
- AttractionLocate trip -> plan
