+++ src/components/plan/AttractionLocate.vue
@@ -8,6 +8,18 @@ const selectGugun = ref("");
 const sidoList = ref([]);
 const gugunList = ref([]);
 
+// 관광 테마 카테고리 선택 12:관광지, 14:문화시설, 15:축제공연행사, 25:여행코스, 28:레포츠, 32:숙박, 38:쇼핑, 39:음식점
+const categoryItems = [
+    { code: 12, name: "관광지" },
+    { code: 14, name: "문화시설" },
+    { code: 15, name: "축제공연행사" },
+    { code: 25, name: "여행코스" },
+    { code: 28, name: "레포츠" },
+    { code: 32, name: "숙박" },
+    { code: 38, name: "쇼핑" },
+    { code: 39, name: "음식점" },
+];
+
 const getSido = async function () {
     await axios.get("http://localhost/trip/sido").then((response) => {
         sidoList.value = response.data.items;
@@ -54,7 +66,9 @@ const getGugun = async function (sido) {
         </select>
         <select name="content">
             <option selected>관광 타입 선택</option>
-            <!-- <option v-for="gugun in gugunList">{{ gugun.gugunName }}</option> -->
+            <option v-for="(category, idx) in categoryItems" :key="idx" :value="category.code">
+                {{ category.name }}
+            </option>
         </select>
     </div>
 </template>
